Add a LICENSE file
Because the skeleton is using the MIT License, we should add the LICENSE
file. It will prevent translators to release the language extension
without a LICENSE file (MIT or not, they will be able to change that
file accoring to their choosen license). This will be explained in the
future L10n documentation.